pyproject.toml
[tool.poetry]
name = "sqlalchemy-dialects-golembase"
version = "0.1.0"
description = "SQLAlchemy dialect for GolemBase database"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "sqlalchemy_dialects_golembase", from = "sqlalchemy_golembase/src" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
golem-base-sdk = "==0.1.0"
SQLAlchemy = ">=1.4.0,<3.0"
golemdb-sql = {path = "./golemdb_sql", develop = true}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-cov = "^5.0"
pytest-asyncio = "^0.23"
ruff = "^0.1.9"
mypy = "^1.8"
pre-commit = "^3.6"
ipython = "^8.12"
jupyter = "^1.0"
requests = "^2.31"


[tool.poetry.plugins."sqlalchemy.dialects"]
golembase = "sqlalchemy_dialects_golembase.dialect:GolemBaseDialect"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["sqlalchemy_golembase/tests", "testapp/tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=sqlalchemy_dialects_golembase --cov-report=html --cov-report=term-missing"

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-third-party = ["sqlalchemy", "fastapi", "pydantic", "pytest", "golem_base_sdk"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true